refactor suggestions
