+++ packages/ui/components/ui/tx/view/registry.tsx
@@ -4,17 +4,13 @@ import {
   Metadata,
 } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/asset/v1/asset_pb';
 
-export function getStakingTokenMetaData(
+export function getFeeAssetMetadataOrDefault(
   chainId: string,
   assetId: AssetId | undefined,
 ): Metadata | undefined {
   const registryClient = new ChainRegistryClient();
   const registry = registryClient.get(chainId);
+  const feeAssetId = assetId ?? registry.stakingAssetId;
 
-  if (assetId === undefined) {
-    assetId = registry.stakingAssetId;
-  }
-  const StakingTokenMetadata = registry.getMetadata(assetId);
-
-  return StakingTokenMetadata;
+  return registry.getMetadata(feeAssetId);
 }
